core: prevent NPE for RunOnce of a headless VM as non-headless
In LibvirtVmXmlBuilder there is a NPE in writing video device to xml which occurs in
case of running a headless VM with Run Once and set a graphic console for that
running (i.e. running this VM for the 1st time as non-headless).

The reason is that the "temporary" VIDEO device for that run
doesn't include specParams and address parameters since it is not
mandatory, so we can't assume it is always set.

Change-Id: I4393a68b2b55c5b1a45ecc3d286a9bf5967aa304
Signed-off-by: Sharon Gratch <sgratch@redhat.com>