core:setting operator role for fencing in PPC
In order to access the PPC PM agent (iLO3, iLO4, IPMIlan) from ovirt
the user must have either operator or administrator role.

oVirt engine is sending to the PPC PM device a default parameter of
"privlvl=ADMINISTRATOR" , this will prevent a user with a operator
role to access the PM operations supported by oVirt.
In order to enable that, the default setting for PPC PM
(iLO3, iLO4, IPMIlan) should be set to "privlvl=OPERATOR"

Change-Id: Ie7b68fcc5466650c08ee4df57f482ce6cc697e7f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1481197
Signed-off-by: emesika <emesika@redhat.com>