engine: cloud init network configuration - validate user input
Add validation for user input of cloud init net config via the
UI in all commands that add\update a VM or a template VM or on
run VM once. These validations are being added as support for
usage of the Openstack Metadata Protocol in cloud-init is introduced in the engine.

cloud-init documentation:
    http://cloudinit.readthedocs.io/en/latest/topics/network-config.html
Openstack Metadata documentation:
    http://specs.openstack.org/openstack/nova-specs/specs/liberty/implemented/metadata-service-network-info.html

[RFE] Enable configuring IPv6 in VM cloud-init

Bug-Url: https://bugzilla.redhat.com/1330217
Change-Id: I9da4fe9ecd11cc7400bd9750ac14027d88c1e9e2
Signed-off-by: Eitan Raviv <eraviv@redhat.com>