core: Register a VM with disk that is already registered
There is an option using the REST-API to register an attached disk.
We need to keep this behavior for DR scenarios.
Therefore, a verification was added to check whether unregistered VM disks
already exist in the DB.

In case one of the disks is already registered to the DB, the VM
registration will fail.

If the allow_partial flag is true, then the VM registration will succeed
and the existing disks registration will skip.

Change-Id: Iafddbd54280c648d97031704d1b463c1cc3fc6d3
Bug-Url: https://bugzilla.redhat.com/1506484
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>