core: Avoid prepare or teardown image during live create snapshot
In commit 76eaf832 (getQemuImageInfo should not call teardown on live
snapshot), prepare and teardown image weren't supposed to be
called when creating snapshots while the VM is running. To do this, the
code checked whether the VM is starting or up; however, there are use cases
where the VM is, temporary, paused during create snapshot and at that
point, it is neither up nor starting. In this patch, we check if the VM
isn't down to decided if prepare and teardown is required.

Change-Id: Ic5309751a13f82045f2af5532326744eaa29d1e2
Bug-Url: https://bugzilla.redhat.com/1510957
Signed-off-by: Ala Hino <ahino@redhat.com>