core: NPE when adding fence agent from API
When adding a fence agent from the API with a wrong UUID a NPE is
thrown.
This occurs also when updating or removing a fence agent with wrong
UUID from the API.

This patch adds the needed validation for the passed host UUId and fix
also the validation message that is displayed to the user.

Change-Id: Ifd2a1c916a62a55a3fadf24385a3407070db5a31
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1508023
Signed-off-by: emesika <emesika@redhat.com>