packaging: Move Ansible playbooks to ovirt-engine
This patch moves Ansible playbooks which were previsouly stored outside
of ovirt-engine project in different git repository[1].

It turn out to be beeter to maintain the playbooks inside the
ovirt-engine project, so the developers don't need install additional
dependencies to their dev env.

This patch adds two playbooks:
 - ovirt-host-deploy
 - ovirt-host-upgrade

Also two roles which are called from previously mentioned playbooks.

The structure in future for creating a new playbooks should be
following:

  Directory which store playbooks.
  /usr/share/ovirt-engine/playbooks

  Directory which store roles:
  /usr/share/ovirt-engine/playbooks/roles

[1] https://github.com/oVirt/ovirt-ansible/

Change-Id: I0bcc5e70c903ccd05d4fdb04ba18f0b70abe01ed
Signed-off-by: Ondra Machacek <omachace@redhat.com>