restapi: Remove cluster validation on register
As part of the DR site to site implementation, the cluster name has
became part of the entity's OVF therefore there is no need to
obligate the user to specify the cluster when executing the register
operation.

In case the user will register an entity without specify the cluster
id/name, the cluster name from the entity's OVF will be the one which
will be used (unless the register request will also include
cluster mapping).

Change-Id: I3afe28d6fc9c1aa0b17856711f634caebd4540a1
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>