core: Add missing indexes on engine DB
An analysis using pgcluu[1] resulted a handful of missing
indexes on 28 of our foreign keys.

Since our schema is read and update mostly, and tables are small
the expected cost of adding the indexes is negligible.

Running this script[2] against the db will be easier than running a full
report.

[1] pgcluu project - https://github.com/darold/pgcluu
[2] sql script to generate the report - https://gist.github.com/rgolangh/2a195f0822ba6b68daf52f44ef33fda4

Change-Id: I2e948f700e23d3aea83101d59a065c8eecddc1fe
Bug-Url: https://bugzilla.redhat.com/1512412
Signed-off-by: Roy Golan <rgolan@redhat.com>