restapi: Move and Copy disk actions take disk profile as a parameter.
When a disk is moved to a different storage domain, its current disk profile
does not exist there. This patch adds a parameter to the move and copy
commands, so user can set a new disk profile.

Leaving the parameter empty will choose the first profile
with the right permissions.

Change-Id: I03a67aec91cad501a6bd1e357d2ef00554989cbf
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>