core: Register template with disk that is already registered
There is an option using the REST-API to register disk that attached
to a template.
We need to keep this behavior for DR scenarios.
Therefore, a verification was added to check whether unregistered
template disks already exists in the DB.

In case one of the disks is already registered to the DB, the Template
registration will fail.

If the allow_partial flag is true, then the Template registration will
succeed and the existing disks registration will skip.

Change-Id: I98cf0e5b259abdcd62e787e2fce58b37fb6317d3
Bug-Url: https://bugzilla.redhat.com/1506484
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>