core: allow key=key=val format in PM options
In some circumstances we have to send for an agent a key that it value
has a key=value format.
For example :
    ssh_options=-oCiphers=+3des-cbc
The fence agent options field supports up till now only a key=value
format.
This patch allows to have in addition to key=value format also
key=key=value format, however this is limited to one key=value pair as
the value of the fence agent key since supporting multiple key=value
pairs as the value of the fence agent key is very complex.

This patch removes redundant validations from the options field.
The responsibility of the administrator is to write the right options
supported by the fence agent and to verify it from the fence agent
script first.
If anything is wrong in the options field, in case that it is a wrong
key it is simply ignored, in case that it is a wrong value, the engine
will display the message that the fence agent script returns.

Change-Id: I76b0238b212461fe878a2d0c2114deb8a8e3a95d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1484404
Signed-off-by: emesika <emesika@redhat.com>