backend: call GetImageTicket in TransferImageCommand
GetImageTicket used to be called from TransferImageStatusCommand, thus
creating two places where an image transfer was updated with the up to
date information from the image ticket for download and upload flows.
In order to maintain a single place where the transfer is updated in
both flows, this patch moves the call to GetImageTicket to
TransferImageCommand and updates the image transfer there.

Change-Id: I86323a4c59212b5af6468ae9ab591ae9460bb82c
Signed-off-by: Idan Shaby <ishaby@redhat.com>