core: pass correlation from MoveDiskCommand to LiveMigrateDiskCommand
When LiveMigrateDiskCommand starts it gets a new correlation ID,
unrelated to MoveDiskCommand which makes it difficult to follow up
on the status of live storage migration if it's invoked externally
(e.g. via the API).

This patch passes MoveDiskCommand's correlation ID to
LiveMigrateDiskCommand to make following up on it simple.

Change-Id: I213cddb73e8328639750918942a6f008807f0f5b
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>