core: increase PG work_mem to 8192kb
Motivation:
PG work_mem is the memory workspace pg uses for hashes/sorts while
executing queries. when the limit is hit it moves to temp_files which is
obviously less desirable.

A system with 30 hosts, ~700 vms and 1300 disks will get hit by this and
with our current scheme could lead to generating several Gb of files per
hour. The queries needing most of those temp files are the audit_log and
dashboard queries.

Using pgbadger to analyze the pg logs, we can see a break down of few
hours:
Temporary Files Activity

Temp Files

Day 	Hour 	Count 	Total size 	Average size
Oct 31	15	660	5.76 GiB	8.94 MiB
	16	1,418	11.70 GiB	8.45 MiB
	17	715	2.33 GiB	3.34 MiB

Modification:
Default the pg setting work_mem to 8192KB or above.

Results:
Depending on the size of the system, most of the queries would not need
any temp files, and the largest queries would hit the work_mem.

Note: The impact of this settings is that each connection can allocate
this amount of memory to perform the query.
Before we go on and increase this number to make sure any
installation size would fit into the work_mem we first optimize the
queries to consume less memory(is sorting needed, is merging needed, do
we have the right indexes to work on the proper dataset size etc)

Change-Id: I4e94d17a250e5c50a3250e8dc8c0df32fcedd8ef
Bug-Url: https://bugzilla.redhat.com/1510432
Signed-off-by: Roy Golan <rgolan@redhat.com>