backend: maintain a single ticket per transfer
When pausing an image transfer, the ticket was used to be removed and a
new one was used to be created upon resuming.
This design complicated the way we handled image transfers, and
therefore it is replaced in this patch.
Now, when pausing a transfer, the ticket is extended (if necessary) and
when the transfer resumes, only the transfer phase is changed.
This way we can manage a single ticket per transfer and make things
easier to maintain.

Change-Id: Ifc75e0f4bd762500cb0e5e7fde037e1c14ca1b60
Bug-Url: https://bugzilla.redhat.com/1513969
Bug-Url: https://bugzilla.redhat.com/1514887
Signed-off-by: Idan Shaby <ishaby@redhat.com>