core: better handling of migrating vm disappearing from source
We might miss the Down event on power-off VM (e.g., since the
monitoring is processing a statistics cycle) and so the monitoring may
notice that this VM disappeared from the source host.

We used to hand-over the VM to the destination in that case. This is
alright in most of the cases since we will then monitor the destination
host and make a safer decision based on that. However, in some scenarios
migrating_to_vds may be null and so we will have a VM with run_on_vds
set to null. We should minimize those cases.

Thus, this patch changes the handling of this scenario. We hand-over
the VM only when migrating_to_vds is not null, i.e., we know what is
the destination host. Otherwise, we treat this as a failure and abort
the migration.

Need to think what to do regarding cold-reboot and auto-start since
it may be too risky to start the VM in this case.

Change-Id: I681021d1ff255d03078405edc37d801b0f18ba02
Bug-Url: https://bugzilla.redhat.com/1505399
Signed-off-by: ahadas@redhat.com <ahadas@redhat.com>