webadmin: set bytesEndOffset for removed upload handlers
When uploading an image via the webadmin, if the browser is refreshed
then the UploadImageHandler is removed. When the engine detects that the
image transfer entity in the db was not updated for some time, the phase
is switched to paused by system. Then, when the user manually resumes
the upload, a new UploadImageHandler is created but its bytesEndOffset
field is not pulled from the db. This leads to the inability of the
handler to continue and upload the rest of the file.
This patch updates this field upon resuming.

Change-Id: I50c4d5ac1966fb8edd2a77191390023102aac73f
Bug-Url: https://bugzilla.redhat.com/1515852
Signed-off-by: Idan Shaby <ishaby@redhat.com>