webadmin: Reflect cluster's default network provider on reinstall
On reinstalling  a host, initially the network provider is set to the
cluster's default network provider. If the user wants to choose another
network provider, he has to disable a checkbox to be able to select
one of the other network providers.

The user might want to override the default network provider, if she
wants a provider configuration on the host which is different from the
one oVirt is configuring automatically. A plausible motivation to do so
is if the user wants to migrate from one provider to another, but do so
gradually.

Change-Id: I51d299c7dc068591937d8a3096a84a8f4a37f668
Signed-off-by: Dominik Holler <dholler@redhat.com>