webadmin: Show neutron agent properties only for neutron
Use the type of the selected network provider to decide if the
neutron agent properties are shown.

Recent changes introduced a plugin type for all external network
providers. For this reason it is not suitable anymore to check if
the network provider has a plugin type to decide if the selected
provider is an OpenStack external network provider and show the
neutron agent properties.

Change-Id: I0ce14fca850e3bd96c0858a43182475e3582fcfe
Signed-off-by: Dominik Holler <dholler@redhat.com>