core: Don't use EngineError.VM_INVALID_SERVER_CLUSTER_ID
EngineErrors are supposed to be used for failed executions, not for
failed command validations.
This patch replaces the single (wrong!) usage of
EngineError.VM_INVALID_SERVER_CLUSTER_ID, and replaces it with the
corresponding EngineMessage with the same meaning,
EngineMessage.ACTION_TYPE_FAILED_CLUSTER_IS_NOT_VALID.

Besides being technically more correct, this patch should also solve
the fact that the failure message for attempting to add a VM with a
non-existing cluster was not I18Ned.

Change-Id: Icf5262f795945c4f306bd67f2bd9f3050eb9adb1
Signed-off-by: Allon Mureinik <amureini@redhat.com>