core: drop the error status of hosts
We used to set the status of a host to Error when we fail to
run VMs on that host that successfully run on other hosts.
This status seems redundant now as we believe we have a good
coverage of the problems a host might have that prevent it
from running VMs (those issues should cause the engine to
mark that host as non-opreational. Since we also encountered
issues with having the Error status, we now remove the only
place that sets it. Later, we will remove this status
together with all the places that use it.

Change-Id: I7f7091b4a767b4aa0a96f110274ee683e1494cdc
Bug-Url: https://bugzilla.redhat.com/1376754
Signed-off-by: ahadas@redhat.com <ahadas@redhat.com>