core: Rename ovn_tunneling_network to ovn_tunneling_interface
The variable ovn_tunneling_network is renamed to
ovn_tunneling_interface, because it stores not only
a network name, but may also store an IP address.

The old alias host_deploy_ovn_tunneling_network was kept for
temporary backward compatibility in ovirt-provider-ovn-driver.
This is not required anymore, because ovirt-provider-ovn-driver
is now in the same repository as InstallVdsInternalCommand.

Change-Id: Ic4c2a7dcc478cae27a4d1a15215b1bf07fcfd944
Signed-off-by: Dominik Holler <dholler@redhat.com>