core, webadmin: Unmanaged network provider proxy.
As part of our integration with ACI, we need to provide external
networks that will later be managed by ACI via a VDSM hook, i.e.
external networks without an external provider we are to manage
(but rather ACI)

This patch introduces an unmanaged proxy (which does not communicate
with an external source) for external networks.

Change-Id: I87f8c6ab43772a88b252877b9f4d21665428653c
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>