core: Update default network provider of cluster
If the default network provider of a cluster is changed,
the new provider is set in the db for all host of the cluster.
The hosts changed in the db are marked as ReinstallRequired,
because the change in the db is not yet applied to the physical
hosts.

The ReinstallRequired is set in UpdateClusterCommand instead of
in the UpdateVdsCommand, because in a follow up patch, the change
of the physical hosts will be applied on cluster level.

Change-Id: I8efd4bb5ef1323698d468bf5f38a4d77187e6858
Bug-Url: https://bugzilla.redhat.com/1528297
Signed-off-by: Dominik Holler <dholler@redhat.com>