core: Add a verification for OVF store disks creation
In change - 7575199d1b7a317c471c93bda42e1ba0c053824b
a verification to all the internal commands in
DeactivateStorageDomainWithOvfUpdateCommand was added.

As part of that change, there should be a verification
whether the internal command to create OVF store disks
succeeded as expected when deactivating a storage domain.

It was missing and cause a storage domain to remain at
maintenance state in case the disks didn't create as expected.

This patch adds the required verification.

Change-Id: I0ee012131c628d6950b64af19afba6c0bc2ce928
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>