core, webadmin: Move import logic from webadmin to backend
A new command is created to hold the logical flow of importing
an external network. This way the functionality can be shared
between webadmin and backend.

The import logic is changed, so that there are no network filters
defined for external networks, because network filters for
Open vSwitch are not supported by libvirt.

Change-Id: I84173ba4887946e07f3d00f528011c628bfcf68b
Signed-off-by: Dominik Holler <dholler@redhat.com>