core: Add end procedure when running DestroyImageCommand
Not setting that end procedure as COMMAND_MANAGED may cause the command
to be cleaned from that data too early, causing CoCo to fail ending the
command appropriately.

This patch is different than the master patch because for
DestroyImageCheckCommand, that is not an async command, the end
procedure should be PARENT_MANAGED.
On the master this change isn't relevant because
DestroyImageCheckCommand was removed.

Change-Id: Ifc964ef7e451baba4fc56461da6e704bb2b67423
Bug-Url: https://bugzilla.redhat.com/1509629
Signed-off-by: Ala hino <ahino@redhat.com>