db: use PG UUID generation function
We have our own implementation of uuid_generate_v1() function, which was introduced
to allow smoother transition between PostgreSQL 8.x and 9.x. But since PG 9.x there
is a standard implementation of uuid_generate_v1() which is provided by uuid-ossp
PostgreSQL extension.

This extension included in postgresql-contrib package which is part of both standard
and SCL versions of PostgreSQL and it's significantly faster than our own implementation.
Unfortunately installing this extension requires administrator privileges, so it cannot
be part of our normal database upgrade scripts, but switching to SCL PostgreSQL 9.5 is
a great opportunity to make this change.

Change-Id: Iea841cbac3e53e0ab598c4b1e23df489df5eabc1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1515635
Signed-off-by: emesika <emesika@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>