engine: Multipath alerts support
In this patch, the multipath health report received as part of
the Host.GetStats polling is used to create AuditLogs for
multipath devices.

The alerts are per host, and a hash mechanism is added in order
to create alerts only if the report change.

The following alarms are added:

No faulty multipath paths on host ${VdsName} :
 - All the multpath devices are functioning on this host.

Faulty multipath paths on host ${VdsName} on devices: [${MpathGuids}]:
 - There are multipath devices on the host with malfunctioning paths.

Multipath devices without valid paths on host ${VdsName} : [${MpathGuids}]:
- There are multipath devices on the host without any functioning paths.

This patch is part of the multipath event feature:
https://www.ovirt.org/develop/release-management/features/storage/multipath-events/

Change-Id: I7c9b05ebc001b6eb6ecac54f2ca2cabc5848dbee
Signed-off-by: Fred Rolland <frolland@redhat.com>