core: Change lock scope to 'Command' in UpdateVMCommand
Since patch facac922dc39011b765ea74e522371a1a4d5a2cf,
RemoveLeaseCommand and AddLeaseCommand become async
and take a while to complete.

The lock scope for UpdateVMCommand is 'Execution' which
applied only to the sync execution part of each command
and therefore new updates can be done during the
execution of creating or removing leases.

Change the scope of the lock to 'Command' and adding the tasks
of adding or removing lease to UpdateVMCommand TaskIdList will
prevent from running updates for the VM until all the update
tasks will end.

Change-Id: I7b027dab6ae25532d98b7c02fbe9a2cc5cd360a4
Bug-Url: https://bugzilla.redhat.com/1516907
Bug-Url: https://bugzilla.redhat.com/1492473
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>