core: Moving or copying disk uses default quota
When moving or copying a disk using REST API
and no new quota is specified, the same quota is used
for the new disk. This is a problem, if the disk has null quota.

This patch changes it so the new disk uses default quota
if the old quota is null.

Change-Id: Id0d4ce2722c77ce4ebf90c413d142084c946ff0a
Bug-Url: https://bugzilla.redhat.com/1518693
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>