core: import ovirt's ova from path
This patch extends ImportVmFromOvaCommand with the ability to
import OVA files that were generated by oVirt from a given
path on a selected host.

If the OVA is detected as one that was generated by oVirt (for
now the check is based on the VM's origin field, but should be
changed to examine the 'ovirt' namespace within the OVF) then
the ansible playbook for extracting disks in an OVA file into
images on storage domain is executed rather than the verb in
VDSM that executes virt-v2v.

Change-Id: Icb5014846bfb05d5dea939f725926043defd353a
Bug-Url: https://bugzilla.redhat.com/1049604
Signed-off-by: ahadas@redhat.com <ahadas@redhat.com>