engine: Allow Adding vNicProfile and Network without user
If no user is set for the AddNetworkCommand and AddVnicProfileCommand,
no user will be defined for the role NETWORK_ADMIN on the added Network
or VnicProfile. Other permissions of the added network or vNIC profile
are not affected. This means that users with UserRole are still able to
use VMs associated to the vNIC profile and SuperUsers and NetworkAdmins
are still able to manage the network and vNIC profile.

This change is required to import external networks from background jobs,
because the background job is not associated to an user.

Change-Id: I51b2d8765d55a17e68f9059e630086c406b82b71
Signed-off-by: Dominik Holler <dholler@redhat.com>