ansible: query ovf configuration from ova files
This patch adds a playbook for gathering the OVF configuration from
an OVA file. With this, we can deprecate the Host.getConvertedVm
verb in VDSM - that will reduce code duplication and will replace
the minimal parsing that was implemented in VDSM with a more
complete one.

The role that this playbook uses extracts the OVF configuration
from the OVA file using a python script. It is implemented that
way since according to the specification the OVF file is the
first entry in the OVF tar file. Its name might change. I
found no way to extract just the first entry of a tar file, and
even if there is - that way, we will need to extract it somewhere
and then read it. The python script does it in a more elegant way.

Change-Id: Idf78aa3a816cf76ea0a06e9642631e7ad1763b01
Bug-Url: https://bugzilla.redhat.com/1049604
Signed-off-by: ahadas@redhat.com <ahadas@redhat.com>