core: Add vlaidation for missing cluster on import
The import command miss several basic validations which should be validated
at the start of the operation.
We should validate if the entity exists as well as the existence of
the cluster and if the cluster storage pool id is matching the storage
pool where the source storage domain exists.

Change-Id: I582c6edb231057c3aa82af42c8b565ccdbdfc2c5
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>