dbscripts: Add constraint to ensure unique external networks
The added constraint ensures, that every external network is imported
only once in data center.

The corresponding validation is already active in
AddNetworkCommand::externalNetworkNewInDataCenter.

It should not happen, that the database does not fulfill the
constraint. But if the database does not fulfill the constraint
during upgrade, the database is considered as corrupted and the
upgrade should abort. In this case the database has to be fixed
manually.

Change-Id: Ia5b9e060e1140e3bd77f6c18a18a135b17ffd4a8
Signed-off-by: Dominik Holler <dholler@redhat.com>