core: Improve logging for failed VM live migration
When migrating a VM with a direct LUN to a host which is not
visible to it, the error logging is insufficient.
The log is missing the reason for the failure which is the LUN
visibility to the host.

The information on which LUN is not visible to the host was added.

Change-Id: I5e6445a4581003bf75c570c74c1451498ef7879b
Bug-Url: https://bugzilla.redhat.com/1457087
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>