core: Extend compensation use to callbacks
Motivation: use the compensation mechanism after the
'sync' execution phase of a command when the command uses
CoCo mechanism.

Currently, when the execution part of a command ends
succssesfully a cleanup for the compensation occurs.
Therefore there is no way to use the compensation object
when the 'sync' part of a command succeeds but one of the
child commands failed.

Added an annotaion variable to determind if the compensation
will stay 'alive' until the end of the executions of all the
child commands or will removed at the end of the execution
part of the command.

Change-Id: Ic9badd5e64c58dedf7d9057390e62234db7e66ed
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>