core: prevent NPE when internal command renames an entity
The audit log that is produced by commands that rename entities,
such as update-VM, relies on having the user that initiated the operation.
However, when the operation is postponed, e.g., VM next-run confiugration
is set, the user that initiated the operation is not available when the
operation is actually done. This patch adds another audit log for this
scenario.

Change-Id: I02da85077468f4d31c2cc6243a3f2f511fb8db20
Bug-Url: https://bugzilla.redhat.com/1524172
Signed-off-by: ahadas@redhat.com <ahadas@redhat.com>