core: VM template disk size caching
VmTemplateComparerByDiskSize has an over-complicated caching
mechanism.
First, putting a value in a WeakHashMap will offer no guarantee it
will still be there when you attempt to retrieve it.
Second, the entire map will be eligible for collection once the
Comparator is, which is questionable, since it stands to reason that
if the templates are sorted by the disk size, this size is used
shortly after/beforehand, meaning it would be calculated more than
once.

This patch aligns the implementation of VMTemaplte#getActualDiskSize
with the more commonly used VM#getActualDiskWithSnapshotsSize by
adding a double member to the template and using it to cache the size
calculation, thus gaining a slight performance improvement in the
cases detailed above and sacrificing only the size of a double in
others.

Change-Id: Id67da1b7562c5db849cc7ee3c5b0e24a9155e7f0
Signed-off-by: Allon Mureinik <amureini@redhat.com>