core: Add a verification for OVF store update step
In change - 7575199d1b7a317c471c93bda42e1ba0c053824b
a verification to all the internal commands in
DeactivateStorageDomainWithOvfUpdateCommand was added.

As part of that, there should be a verification
whether the internal command to update OVF store disks
succeeded as expected when deactivating a storage domain.

It was missing and cause to the deactivation of the storage
domain to succeed even if the update failed.

This patch adds the required verification.

Change-Id: I3585d20127b116633351746292058a04f948365e
Bug-Url: https://bugzilla.redhat.com/1523225
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>