ansible: enable logging for ovirt-ova-query
Leverage the new option that was added to the infrastructure for
ansible invocations that enables ordinary logging together with
receiving the output of the playbook in the engine.

In addition, adding a callback-plugin to the ovirt-ova-query
role that reports the OVF in a simpler way that saves the engine
from doing ugly parsing.

Lastly, changing the python script that is executed by this role
to exit with exit-code 2 when the input is invalid.

Change-Id: I1667b987976eae163d1f9bf62029223204e164d5
Bug-Url: https://bugzilla.redhat.com/1049604
Signed-off-by: ahadas@redhat.com <ahadas@redhat.com>