core: Change AsyncTaskDao.getAsyncTaskIdsByStoragePoolId return value
Change AsyncTaskDao.getAsyncTaskIdsByStoragePoolId to return
List<AsyncTask> instead of List<Guid>. This change is required by the
subsequent patch, to log the details of the running async tasks. This change
doesn't affect existing consumers as they only check whether
the size whether the of the returned list is greater than zero.

Change-Id: Ie69d01f0b39055ca601c1ab5dfebd14266153de8
Bug-Url: https://bugzilla.redhat.com/1522669
Signed-off-by: Ala hino <ahino@redhat.com>