setup: add a warn to setup if images with zero 'vm_snapshot_id' exist
Due to a very old v2v bug, there might be images in DB with
'vm_snapshot_id' field set to '00000000-0000-0000-0000-000000000000' in
images table.

In case there is an image in DB with 'vm_snapshot_id' field in the
images table set to '00000000-0000-0000-0000-000000000000' value, the
behavior of the VM assigned with this invalid image may be inconsistent,
E.g: removing the snapshot doesn't remove the invalid snapshot's image,
the invalid image is not displayed in snapshots tab as it should,
commmit to this invalid snapshot failed, clone from this invalid
snapshot failed, running the vm which is assigned with the invalid
snapshot as stateless, caused running failures, etc.

This patch helps the user to detect this invalid images by detecting 3
scenarios described in BZ for this case of zeros in 'vm_snapshot_id'
field and add a warning to engine-setup with a message describing the
invalid VM names, disk aliases and images ids, recommending the user
to consult support for resolving and warn it that the upgrade
will be blocked in the subsequent release (4.3) if not resolved.

The 3 detected scenarios for invalid images are:
1. The 'vm_snapshot_id' field in the images table set to
'00000000-0000-0000-0000-000000000000' value.

2. The 'vm_snapshot_id' field in the images table set to
'00000000-0000-0000-0000-000000000000' value and the corresponding
snapshot entry for this image in snapshots table was removed.

3. The 'vm_snapshot_id' field in the images table set to
'00000000-0000-0000-0000-000000000000' value and the corresponding
snapshot entry for this image in snapshots table is set with 'snapshot_id'
field set also to '00000000-0000-0000-0000-000000000000' value.

Change-Id: Ieb43fb6d3c7c57310eabe42948f9d3fe6cbbb526
Bug-Url: https://bugzilla.redhat.com/1492838
Signed-off-by: Sharon Gratch <sgratch@redhat.com>