packaging: setup: postgres95: Fixes
- Upgrade in main transaction

Before this patch, postgres upgrade was done in its own transaction.
So rollback could only happen if there was a failure in the upgrade
process.

Move this to otopi's MAIN_TRANSACTION, which is where most changes are
done - so that if we fail elsewhere and rollback, we also rollback
postgresql.

Depending on the failure point, the target dir could still not exists.
Ignoring errors on rmdir on abort.

- Do not allow in-place upgrade

It was decided that this is too risky. Most of the code still left in,
only removing the question.

- Do not backup/restore on upgrade

Not needed, because we can revert to the old postgresql for rollback.

Change-Id: I5b4430bfe79df9a4c0e0ffc1696c821e20584eb3
Bug-Url: https://bugzilla.redhat.com/1492138
Signed-off-by: Yedidyah Bar David <didi@redhat.com>