core: ImagesHandler#prepareSnapshotConfigWithAlternateImage
This patch supports null snapshots being passed to
ImagesHandler#prepareSnapshotConfigWithAlternateImage. While there
shouldn't be any valid flow that passes a null snapshot to this
method, past bugs may have left users' database with such
inconsistent states, as noted in the bug below.

This patch does NOT solve all the flows that may be affected by such
a corrupted database, but it solve removing a disk in such a
situation. For a null snapshot, it will just return null, allowing
the following code to fail with a NullPointerException if it doesn't
test such a situation instead of failing with it itself.

Change-Id: I423f8a0b854134eecb9af4f648b7d79f84cc6d77
Bug-Url: https://bugzilla.redhat.com/1527134
Signed-off-by: Allon Mureinik <amureini@redhat.com>