webadmin: Change the lines' color on iSCSI LUNs table
Screenshot: https://imgur.com/a/7LTH7

If a LUN is included on current SD, its line should be black.
By pressing the "Add" button, the line turns from gray to dark-blue.
For all other cases, line is grayed out.
When removing LUNs, lines are grayed out and the V icon is being faded.

Change-Id: I4255815fdc33e7859b10d67de80e38b257a430c3
Signed-off-by: Shani Leviim <sleviim@redhat.com>