webadmin: focus Discard After Delete when it's invalid
When creating/editing a block storage domain, if one of the selected
LUNs doesn't support discard, and the Discard After Delete (DAD)
checkbox is checked, then the model doesn't pass validation and the DAD
checkbox is marked as problematic.
The problem is that DAD is located under the Advanced Parameteres
expander, which is not expanded in this a case.

This patch expands this expander and focuses on the DAD checkbox if it's
invalid, to notify the user why he cannot create/edit the storage
domain.

Change-Id: I9dfc64bd9741bc2b031a2441929c16b6b22d17b4
Bug-Url: https://bugzilla.redhat.com/1502696
Signed-off-by: Idan Shaby <ishaby@redhat.com>