core: Inline DiskByBootAndSnapshotComparator
DiskByBootAndSnapshotComparator is a used in a single place and is
extremely specific to the VmInfoBuildUtils' logic.
This patch inlines it to make the logic clear and to remove some
bloating from the codebase.

Note that the logic here could probably be simplified, and the sorting
according to snapshots could probably be applied to all the disks as a
secondary sorting, regardless of whether they're bootable or not, but
that change is out of scope for this patch, that just focuses on
cleaning up the code, not on the logic.

Change-Id: I7047ad59c3ec2c395fa9c2ddac33620afb7d0aa2
Signed-off-by: Allon Mureinik <amureini@redhat.com>