core: Remove double sorting in VmInfoBuildUtils
VmInfoBuildUtils#getSortedDisks(VM) makes two consecutive sort calls:
The first according to the disk's name and then another sort
according to place the bootable disk(s) at the beginning of the list.

Since sorts in Java are stable and do not change the order of
elements which are equivalent (i.e., in this case, disks that have
previously been sorted by their name won't be resorted if both are
non-bootable) this works as intended, but it's very unintuitive and
hard to read.
A cleaner implementation would be to sort the list once, with a
compound comparator which is first based on the boot flag and only
then by the disk's name.
As a side bonus, this reimplementation should also slightly improve
the performance of this method, since the list is only sorted once.

Change-Id: Ibb41d8df0e45cabc88443c224b9333a4bb778a67
Signed-off-by: Allon Mureinik <amureini@redhat.com>