core: improve the robustness of hosted-engine disks parsing
Recent changes in VDSM in the context of engine XML changed the
output of FullList. It no longer reports the actual size, virtual
size and the device-id of disks and thus import of an hosted
engine VM fails. This patch makes the parsing of the disks of an
hosted engine VM more robust to support this.

Change-Id: I1e107c7feec685c864caa7e2744f7b31e5ffb1e1
Bug-Url: https://bugzilla.redhat.com/1527318
Signed-off-by: Arik Hadas <ahadas@redhat.com>