core: add 'shared' parameter to disk metadata in domain XML
Since compatibility version 4.2 the engine composes the VM XML that
passed to libvirt.
As part of this change, the XML should contain a field to indicate
whether the VM disk device is a transient (backup) disk or not.

Add 'shared' field to VM disk device metadata in the domain XML.
The 'shared' property will indicate whether the VM disk device is
transient, and if so, the VDSM will create the transient
volume under /var/lib/vdsm/transient on the relevant host.

Change-Id: I33e1d0285314d1dc5c28769d435ec1860768dd66
Bug-Url: https://bugzilla.redhat.com/1526815
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>