Revert "core: ansible: Don't print errors when files not found"
Verified this was the cause of error found on OST:
http://jenkins.ovirt.org/job/ovirt-master_change-queue-tester/4520/

Possible cause ( in the comments ):
The full path to the rules file should be specified here. 
"{{ role_path }}/vars/{{ firewald_rules_file }} should work.

Unlike the "include_vars" module, the stat module isn't 
looking for the files inside the "vars" dir of the role.

This reverts commit 1d2825002fb0948b162e1caf695ad2a88c88155f.

Change-Id: I0adb4c716af2483a7d81204159be5fdf1018e76d