Update dependencies to match Wildfly 11
This patch updates the version of some dependencies in
ovirt-engine to match the versions provided at runtime
by Wildfly.

The most correct approach would be to import WildFly's
Bill-Of-Materials (BOM), but due to some technical
difficulties this approach has been suspended for the
time being.

Two specific dependencies have not been upgraded:

1) hibernate-validator

ovirt-engine depends on 4.2.0-Final, Wildfly provides 5.3.5-Final.
Updating the version causes multiple engine tests to fail.
Fixing these tests and updating the version should be done
in a separate pach. This should be revisited.

2) slf4j-api
ovirt-engine depends on 1.7.5, Wildfly provides 1.7.22
This artifact is used extensively, including by some tools
which won't be run under Wildfly, so I preferred not to
make a change here

Change-Id: I092359f6918c75806fb809d672d081936ee8e018
Bug-Url: https://bugzilla.redhat.com/1406766
Signed-off-by: Ori Liel <oliel@redhat.com>