core: do not assume ovf to be the first entry of ova
We relied on the OVF specification that makes it clear that the
OVF configuration should be the first entry of an OVA tar file.
However, apparently some other virtualization management systems
do not obey to this standard and put the OVF configuration
elsewhere inside the OVA file. With this patch, we are going
to support such OVA files (the specification says that the OVF
entry in OVA files must have the 'ovf' file extension).

Change-Id: If32217ac9ca7be2b14832608dfd656da95039028
Bug-Url: https://bugzilla.redhat.com/1528724
Signed-off-by: Arik Hadas <ahadas@redhat.com>