ansible: support for an external ova in the form of a directory
The OVF specification says that OVA can be a directory or a tar file.
The recent switch to query OVF configuration within OVAs using Ansible
broke this ability since the assumption was that the OVA is in the form
of a tar file only. This patch enhances the new mechanism with support
for OVA in the form of a directory. For now, importing OVA that is in
the form of a directory is possible only for those that are generated
by others (e.g., VMware).

Change-Id: I038df3af9c11ccd56d763762f5c8e512d4410036
Bug-Url: https://bugzilla.redhat.com/1529255
Signed-off-by: Arik Hadas <ahadas@redhat.com>