ansible: fix extraction of disks not aligned with 8mb blocks from ova
The disks inside an OVA tar file may not be aligned with the size
of the buffer we use (currently 8mb) and thus the last read of
such a disk includes data that belongs to the next tar entry.
In this case, we need to move the current position inside the
tar file back to the point where the next tar entry starts.

Change-Id: I6d593d26605130354a2b3f829994b902b7571d0d
Signed-off-by: Arik Hadas <ahadas@redhat.com>