core: extract disks matching logic
Extract the current logic used for correlating reported disk
devices with the ones in the database to a separate method.
In a later version of 4.2, we will hopefully be able to replace
it with simpler logic that correlates them according to the
device's alias (we need Libvirt to respect aliases provided
by ovirt-engine for that).

Change-Id: I0e6b02d4a7a9e7a95860048d94890e81dcb11195
Signed-off-by: Arik Hadas <ahadas@redhat.com>