core: improve logging on failure to run vm
Generally, we print the stacktrace only on DEBUG-level logging, as
we do not want to dump too much data into the log file. However, in
light of the recent engine-XML changes and the fact that failures
on run-VM are not that frequent, we can print the stacktrace in
ERROR-level logging to improve our debug capabilities.

Change-Id: I77a869e5a56a4f738c95d4091cab5ccf70323587
Signed-off-by: Arik Hadas <ahadas@redhat.com>