common: use ping2, confirmConnectivity verbs
For cluster compatibility >= 4.2 Host.ping2 verb replaces Host.ping
verb when engine polls the host to ascertain that vdsm is reachable.
A new verb Host.confirmConnectivity is introduced when engine polls
the host to allow vdsm to ascertain that the engine is reachable,
e.g. following network update\setup.
This patch adds support for the new verbs per the cluster compatiblity
level.

Change-Id: I0d05a4b9e53024fbe8cadc272f46aa2d33dd8d6d
Bug-Url: https://bugzilla.redhat.com/1525637
Signed-off-by: Eitan Raviv <eraviv@redhat.com>